DEV-44 Changing PR suggestions
