+++ src/components/MessageLogger/MessageList.module.scss
@@ -1,3 +1,5 @@
+@use "@styles/overrideDefaultMethods.scss";
+
 #containerMessages {
     background-color: black;
     width: 35rem;
@@ -8,7 +10,6 @@
 
 .lineMsgUl {
     width: 100%;
-    height: 100%;
     padding-top: 0.3rem;
-    padding-left: 0;
+    @include overrideDefaultMethods.override-ul;
 }
