+++ src/components/MessageLogger/consoleList.tsx
@@ -1,5 +1,5 @@
-import React, { useState } from "react";
-import { Message } from "@components/MessageLogger/structs/Message";
+import React, { useEffect, useState } from "react";
+import { Message, MessageCounter } from "@components/MessageLogger/structs/Message";
 import { MessageList } from "@components/MessageLogger/MessageList";
 
 interface Props {
@@ -8,13 +8,84 @@ interface Props {
 }
 
 export const ConsoleList = ({ title, messages }: Props) => {
+    const [messagesWithCounts, setMessagesWithCounts] = useState([] as MessageCounter[]);
+
+    useEffect(() => {
+            let contadores: number[] = messagesRepeated(messages)
+            createMessagesWithCounts(contadores)
+    }, [])
+
+
+
+    const createMessagesWithCounts = (contadores: number[]): void => {
+        let items = [] as MessageCounter[];
+
+        for(let i = 0, j = 0; i < messages.length && j < contadores.length; i++, j++){
+            let item: MessageCounter = {
+                msg: messages[i],
+                count: contadores[j]
+            }
+
+            items.push(item)
+
+            if(item.count > 1){
+                //several elements mustn't be represented because they are repeated
+                i += item.count-1
+            }
+        }
+
+        setMessagesWithCounts(
+            [
+                ...items
+            ]
+        );
+    }
 
     return (
         <>
             <h2>{title}</h2>
-            <MessageList messages={messages} />
+            <MessageList messages={messagesWithCounts} />
         </>
 
     )
 
+}
+
+const messagesRepeated = (messages: Message[]): (number[]) => {
+    let counts: number[] = []
+    let count: number = 1
+
+    for(let i = 0; i < messages.length; i++){
+        let el = messages[i]
+        if(i < messages.length - 1){
+            count = checkDuplicated(el.id, messages, i + 1)
+        } else{
+            //if it is in the last element it is because is alone
+            count = 1 
+        }
+        
+        counts.push(count)
+        if(count > 1){
+            //several elements mustn't be counted because they are repeated
+            i += count-1
+        }
+    } 
+    return counts
+}
+
+const checkDuplicated = (id: string, messages: Message[], index: number): number => {
+    let count: number = 1
+    let finished: boolean = false
+
+    while(!finished && index < messages.length){
+        if (id === messages[index].id) {
+            count++
+        } else{
+            finished = true
+        }
+
+        index++
+    }
+
+    return count
 }
