+++ src/styles/overrideDefaultMethods.scss
@@ -1,0 +1,3 @@
+@mixin override-ul{
+    padding-left: 0;
+}
